- change blocks folder
