+++ blocks-src/um-forms/block.json
@@ -1,12 +1,12 @@
 {
   "apiVersion": 2,
-  "name": "um-block/um-member-directories",
-  "title": "Member Directory",
-  "description": "Choose display directory",
-  "icon": "groups",
+  "name": "um-block/um-forms",
+  "title": "Form",
+  "description": "Choose display form",
+  "icon": "forms",
   "category": "um-blocks",
   "attributes": {
-    "member_id": {
+    "form_id": {
       "type": "integer"
     }
   },
