+++ includes/blocks/blocks-manifest.php
