+++ includes/blocks/um-member-directories/src/index.js
@@ -1,107 +1,1 @@
-import { useSelect } from '@wordpress/data';
-import { PanelBody, SelectControl, Spinner } from '@wordpress/components';
-import { InspectorControls, useBlockProps } from '@wordpress/block-editor';
-import ServerSideRender from '@wordpress/server-side-render';
-import { registerBlockType } from "@wordpress/blocks";
-import { useMemo } from '@wordpress/element';
-
-registerBlockType('um-block/um-member-directories', {
-	edit: function (props) {
-		const blockProps = useBlockProps();
-		const { attributes, setAttributes } = props;
-		const { member_id } = attributes;
-
-		const posts = useSelect((select) => {
-			return select('core').getEntityRecords('postType', 'um_directory', {
-				per_page: -1,
-				_fields: ['id', 'title']
-			});
-		}, []);
-
-		const options = useMemo(() => {
-			if (!posts) {
-				return [{ label: wp.i18n.__('Loading...', 'ultimate-member'), value: '' }];
-			}
-			if (posts.length === 0) {
-				return [{ label: wp.i18n.__('No posts found.', 'ultimate-member'), value: '' }];
-			}
-			return [{ label: wp.i18n.__('Select Directory', 'ultimate-member'), value: '' }].concat(
-				posts.map((post) => ({
-					label: post.title.rendered,
-					value: post.id
-				}))
-			);
-		}, [posts]);
-
-		const onMemberIdChange = (value) => setAttributes({ member_id: value });
-
-		if (!posts) {
-			return (
-				<p>
-					<Spinner />
-					{wp.i18n.__('Loading...', 'ultimate-member')}
-				</p>
-			);
-		}
-
-		return (
-			<div {...blockProps}>
-				<ServerSideRender block="um-block/um-member-directories" attributes={attributes} />
-				<InspectorControls>
-					<PanelBody title={wp.i18n.__('Select Directories', 'ultimate-member')}>
-						<SelectControl
-							label={wp.i18n.__('Select Directories', 'ultimate-member')}
-							className="um_select_directory"
-							value={member_id}
-							options={options}
-							onChange={onMemberIdChange}
-						/>
-					</PanelBody>
-				</InspectorControls>
-			</div>
-		);
-	},
-
-	save: () => null
-});
-
-jQuery(window).on( 'load', function($) {
-	let observer = new MutationObserver(function(mutations) {
-		mutations.forEach(function(mutation) {
-
-			jQuery(mutation.addedNodes).find('.um.um-directory').each(function() {
-				let wrapper = document.querySelector('.um-directory');
-				let directory = jQuery(this);
-				um_ajax_get_members( directory );
-				um_slider_filter_init( directory );
-				um_datepicker_filter_init( directory );
-				um_timepicker_filter_init( directory );
-
-				if (wrapper) {
-					wrapper.addEventListener('click', (event) => {
-						if (event.target !== wrapper) {
-							event.preventDefault();
-							event.stopPropagation();
-						}
-					});
-				}
-			});
-			jQuery(mutation.addedNodes).find('.um-member').each(function() {
-				let wrapper = document.querySelector('.um-member');
-				let directory = jQuery(this).parent();
-				UM_Member_Grid(directory);
-
-				if (wrapper) {
-					wrapper.addEventListener('click', (event) => {
-						if (event.target !== wrapper) {
-							event.preventDefault();
-							event.stopPropagation();
-						}
-					});
-				}
-			});
-		});
-	});
-
-	observer.observe(document, {attributes: false, childList: true, characterData: false, subtree:true});
-});
+(()=>{"use strict";var e={n:t=>{var r=t&&t.__esModule?()=>t.default:()=>t;return e.d(r,{a:r}),r},d:(t,r)=>{for(var i in r)e.o(r,i)&&!e.o(t,i)&&Object.defineProperty(t,i,{enumerable:!0,get:r[i]})},o:(e,t)=>Object.prototype.hasOwnProperty.call(e,t)};const t=window.wp.data,r=window.wp.components,i=window.wp.blockEditor,n=window.wp.serverSideRender;var o=e.n(n);const l=window.wp.blocks,a=window.wp.element,c=window.ReactJSXRuntime;(0,l.registerBlockType)("um-block/um-member-directories",{edit:function(e){const n=(0,i.useBlockProps)(),{attributes:l,setAttributes:d}=e,{member_id:u}=l,m=(0,t.useSelect)((e=>e("core").getEntityRecords("postType","um_directory",{per_page:-1,_fields:["id","title"]})),[]),s=(0,a.useMemo)((()=>m?0===m.length?[{label:wp.i18n.__("No posts found.","ultimate-member"),value:""}]:[{label:wp.i18n.__("Select Directory","ultimate-member"),value:""}].concat(m.map((e=>({label:e.title.rendered,value:e.id})))):[{label:wp.i18n.__("Loading...","ultimate-member"),value:""}]),[m]);return m?(0,c.jsxs)("div",{...n,children:[(0,c.jsx)(o(),{block:"um-block/um-member-directories",attributes:l}),(0,c.jsx)(i.InspectorControls,{children:(0,c.jsx)(r.PanelBody,{title:wp.i18n.__("Select Directories","ultimate-member"),children:(0,c.jsx)(r.SelectControl,{label:wp.i18n.__("Select Directories","ultimate-member"),className:"um_select_directory",value:u,options:s,onChange:e=>d({member_id:e})})})})]}):(0,c.jsxs)("p",{children:[(0,c.jsx)(r.Spinner,{}),wp.i18n.__("Loading...","ultimate-member")]})},save:()=>null}),jQuery(window).on("load",(function(e){new MutationObserver((function(e){e.forEach((function(e){jQuery(e.addedNodes).find(".um.um-directory").each((function(){let e=document.querySelector(".um-directory"),t=jQuery(this);um_ajax_get_members(t),um_slider_filter_init(t),um_datepicker_filter_init(t),um_timepicker_filter_init(t),e&&e.addEventListener("click",(t=>{t.target!==e&&(t.preventDefault(),t.stopPropagation())}))})),jQuery(e.addedNodes).find(".um-member").each((function(){let e=document.querySelector(".um-member"),t=jQuery(this).parent();UM_Member_Grid(t),e&&e.addEventListener("click",(t=>{t.target!==e&&(t.preventDefault(),t.stopPropagation())}))}))}))})).observe(document,{attributes:!1,childList:!0,characterData:!1,subtree:!0})}))})();
