+++ includes/blocks/um-password-reset/src/index.js
@@ -1,17 +1,1 @@
-import { registerBlockType } from '@wordpress/blocks';
-import ServerSideRender from '@wordpress/server-side-render';
-import { useBlockProps } from '@wordpress/block-editor';
-
-registerBlockType('um-block/um-password-reset', {
-	edit: function () {
-		const blockProps = useBlockProps();
-
-		return (
-			<div {...blockProps}>
-				<ServerSideRender block="um-block/um-password-reset" />
-			</div>
-		);
-	},
-
-	save: () => null
-});
+(()=>{"use strict";var e={n:o=>{var r=o&&o.__esModule?()=>o.default:()=>o;return e.d(r,{a:r}),r},d:(o,r)=>{for(var t in r)e.o(r,t)&&!e.o(o,t)&&Object.defineProperty(o,t,{enumerable:!0,get:r[t]})},o:(e,o)=>Object.prototype.hasOwnProperty.call(e,o)};const o=window.wp.blocks,r=window.wp.serverSideRender;var t=e.n(r);const n=window.wp.blockEditor,s=window.ReactJSXRuntime;(0,o.registerBlockType)("um-block/um-password-reset",{edit:function(){const e=(0,n.useBlockProps)();return(0,s.jsx)("div",{...e,children:(0,s.jsx)(t(),{block:"um-block/um-password-reset"})})},save:()=>null})})();
