+++ includes/core/class-blocks.php
@@ -101,7 +101,7 @@ if ( ! class_exists( 'um\core\Blocks' ) ) {
 			}
 
 			if ( function_exists( 'wp_register_block_metadata_collection' ) ) {
-				wp_register_block_metadata_collection( UM_PATH . 'build', UM_PATH . 'build/blocks-manifest.php' );
+				wp_register_block_metadata_collection( UM_PATH . 'includes/blocks', UM_PATH . 'includes/blocks/blocks-manifest.php' );
 			}
 
 			$blocks = array(
@@ -136,7 +136,7 @@ if ( ! class_exists( 'um\core\Blocks' ) ) {
 
 			foreach ( $blocks as $k => $block_data ) {
 				$block_type = str_replace( 'um-block/', '', $k );
-				register_block_type_from_metadata( UM_PATH . 'build/' . $block_type, $block_data );
+				register_block_type_from_metadata( UM_PATH . 'includes/blocks/' . $block_type, $block_data );
 			}
 		}
